Make nested object deletion extensible
It's currently possible to extend create and update functionality by just implementing those methods on the child serializer, but deletion is a bit trickier.

This proposed refactor exposes just the place in the code where deletion takes place so that more functionality can be easily added in user code if desired.

For my particular use case, I want to be able to create events and update inventory any time a nested item is deleted. With this new change, I can simply extend ``perform_nested_delete(pks_to_delete, **kwargs)`` to include my functionality and call super() to ensure the deletion takes place as expected.

As a side effect, this enables users to adopt functionality proposed in #103 and #112 until the PRs are accepted/declined.